backend/requirements.txt
fastapi
annotated-doc
uvicorn[standard]
python-dotenv
pydantic-settings
mysql-connector-python
bcrypt
python-jose[cryptography]
pandas
openpyxl
httpx
python-multipart
